+++ docs-markdown/src/controllers/cleanup/handleSingleValuedMetadata.ts
@@ -1,9 +1,6 @@
-import { reporter } from "../../helper/telemetry";
 import { handleMarkdownMetadata } from "./handleMarkdownMetadata";
 import { handleYamlMetadata } from "./handleYamlMetadata";
 import { readWriteFileWithProgress } from "./utilities";
-
-const telemetryCommand: string = "applyCleanup";
 
 /**
  * Searches through all directories from rootPath
@@ -11,7 +8,6 @@ const telemetryCommand: string = "applyCleanup";
  * then converts the array to single item.
  */
 export function handleSingleValuedMetadata(progress: any, file: string, files: string[] | null, index: number | null) {
-    reporter.sendTelemetryEvent("command", { command: telemetryCommand });
     const message = "Single-Valued metadata";
     if (file.endsWith(".yml") || file.endsWith(".md")) {
         return readWriteFileWithProgress(progress,