+++ docs-markdown/src/controllers/image-controller.ts
@@ -3,8 +3,8 @@ import Axios from "axios";
 import { existsSync } from "fs";
 import { basename, dirname, extname, join, relative } from "path";
 import * as recursive from "recursive-readdir";
-import { CompletionItem, Position, QuickPickItem, QuickPickOptions, window, workspace } from "vscode";
-import { hasValidWorkSpaceRootPath, ignoreFiles, insertContentToEditor, isMarkdownFileCheck, isValidEditor, noActiveEditorMessage, setCursorPosition } from "../helper/common";
+import { CompletionItem, InputBoxOptions, Position, QuickPickItem, QuickPickOptions, window, workspace } from "vscode";
+import { hasValidWorkSpaceRootPath, ignoreFiles, insertContentToEditor, isMarkdownFileCheck, isValidEditor, noActiveEditorMessage, postWarning, setCursorPosition } from "../helper/common";
 import { sendTelemetryData } from "../helper/telemetry";
 
 const telemetryCommandMedia: string = "insertMedia";
@@ -23,6 +23,7 @@ export function insertImageCommand() {
         { command: applyComplex.name, callback: applyComplex },
         { command: applyLocScope.name, callback: applyLocScope },
         { command: applyLightbox.name, callback: applyLightbox },
+        { command: applyLink.name, callback: applyLink },
     ];
     return commands;
 }
@@ -54,6 +55,10 @@ export function pickImageType() {
         description: "",
         label: "Add lightbox to image",
     });
+    items.push({
+        description: "",
+        label: "Add link to image",
+    });
     window.showQuickPick(items, opts).then((selection) => {
         if (!selection) {
             return;
@@ -62,10 +67,6 @@ export function pickImageType() {
             case "image":
                 applyImage();
                 commandOption = "image";
-                break;
-            case "image":
-                applyImage();
-                commandOption = "image (docs markdown)";
                 break;
             case "icon image":
                 applyIcon();
@@ -82,6 +83,10 @@ export function pickImageType() {
             case "add lightbox to image":
                 applyLightbox();
                 commandOption = "lightbox";
+                break;
+            case "add link to image":
+                applyLink();
+                commandOption = "link";
                 break;
         }
         sendTelemetryData(telemetryCommand, commandOption);
@@ -388,9 +393,8 @@ export async function applyLocScope() {
         noActiveEditorMessage();
         return;
     }
-
     // if user has not selected any text, then continue
-    const RE_LOC_SCOPE = /:::image\s+((source|type|alt-text|lightbox|border)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
+    const RE_LOC_SCOPE = /:::image\s+((source|type|alt-text|lightbox|border|link)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
     const position = new Position(editor.selection.active.line, editor.selection.active.character);
     // get the current editor position and check if user is inside :::image::: tags
     const wordRange = editor.document.getWordRangeAtPosition(position, RE_LOC_SCOPE);
@@ -420,7 +424,7 @@ export async function applyLocScope() {
             });
         }
     } else {
-        const RE_LOC_SCOPE_EXISTS = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
+        const RE_LOC_SCOPE_EXISTS = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope|link)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
         const locScopeAlreadyExists = editor.document.getWordRangeAtPosition(position, RE_LOC_SCOPE_EXISTS);
         if (locScopeAlreadyExists) {
             window.showErrorMessage("loc-scope attribute already exists on :::image::: tag.");
@@ -441,7 +445,7 @@ export async function applyLightbox() {
     }
 
     // if user has not selected any text, then continue
-    const RE_LIGHTBOX = /:::image\s+((source|type|alt-text|loc-scope|border)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
+    const RE_LIGHTBOX = /:::image\s+((source|type|alt-text|loc-scope|border|link)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
     const position = new Position(editor.selection.active.line, editor.selection.active.character);
     // get the current editor position and check if user is inside :::image::: tags
     const wordRange = editor.document.getWordRangeAtPosition(position, RE_LIGHTBOX);
@@ -479,7 +483,7 @@ export async function applyLightbox() {
             }
         });
     } else {
-        const RE_LIGHTBOX_EXISTS = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
+        const RE_LIGHTBOX_EXISTS = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope|link)="([a-zA-Z0-9_.\/ -]+)"\s*)+:::/gm;
         const lightboxAlreadyExists = editor.document.getWordRangeAtPosition(position, RE_LIGHTBOX_EXISTS);
         if (lightboxAlreadyExists) {
             window.showErrorMessage("lightbox attribute already exists on :::image::: tag.");
@@ -507,4 +511,43 @@ export function imageCompletionProvider() {
     completionItems.push(new CompletionItem(`:::image type="icon" source="" alt-text="" loc-scope="":::`));
     completionItems.push(new CompletionItem(`:::image type="complex" source="" alt-text="" loc-scope="":::`));
     return completionItems;
+}
+
+export async function applyLink() {
+    // get editor, its needed to apply the output to editor window.
+    const editor = window.activeTextEditor;
+    if (!editor) {
+        noActiveEditorMessage();
+        return;
+    }
+
+    // if user has not selected any text, then continue
+    const RE_LINK = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope)=""(.*?)"\s*)+:::/gm;
+    const position = new Position(editor.selection.active.line, editor.selection.active.character);
+    // get the current editor position and check if user is inside :::image::: tags
+    const wordRange = editor.document.getWordRangeAtPosition(position, RE_LINK);
+    if (wordRange) {
+        const options: InputBoxOptions = {
+            placeHolder: "Enter link (URL or relative path)",
+        };
+        window.showInputBox(options).then((imageLink) => {
+            if (imageLink === undefined) {
+                postWarning("No link provided, abandoning command.");
+            } else {
+                editor.edit((selected) => {
+                    selected.insert(new Position(wordRange.end.line, wordRange.end.character - 3), ` link="${imageLink}"`);
+                });
+            }
+        });
+    } else {
+        const RE_LINK_EXISTS = /:::image\s+((source|type|alt-text|lightbox|border|loc-scope|link)="(.*?)"\s*)+:::/gm;
+        const linkAlreadyExists = editor.document.getWordRangeAtPosition(position, RE_LINK_EXISTS);
+        if (linkAlreadyExists) {
+            window.showErrorMessage("link attribute already exists on :::image::: tag.");
+            return;
+        }
+
+        window.showErrorMessage("invalid cursor position. You must be inside :::image::: tags.");
+    }
+    return;
 }